Separate StateTreeViewItemModel
